requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
python-multipart
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]